Split catalog.py into extraction, maintenance, and search modules
catalog.py was 1297 lines. Extract three focused modules:
- extraction.py (147 lines): index value extraction from content objects
- maintenance.py (155 lines): standalone PG maintenance ops, compat shims
- search.py (106 lines): core search executor (_run_search)

catalog.py retains the PlonePGCatalogTool class with security declarations,
write/read paths, and ZMI helpers (957 lines, down 26%).
